requirements.txt
psycopg2==2.8.6
python-dotenv==0.15.0
SQLAlchemy==1.3.20
